# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Midanic E-Commerce & ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads/{id}:
    get:
      operationId: getUploadedImage
      tags: [storage]
      summary: Serve a previously uploaded image (public, no auth required)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Image file
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
  /uploads:
    post:
      operationId: uploadImage
      tags: [storage]
      summary: Upload an image file (multipart/form-data, admin auth required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Uploaded image record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedImage"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadedImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        size:
          type: integer
      required: [id, url, objectPath, contentType, size]

    # ── Product Attribute Schemas ────────────────────────────────────
...